# specs/001-b2b-stock-gateway/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: B2B Stock Gateway API
  version: 0.2.0
servers:
  - url: http://localhost:3000
paths:
  /api/v1/products:
    get:
      summary: List products for an authenticated B2B company
      tags:
        - Products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authorized product list
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    required:
                      - source
                    properties:
                      source:
                        type: string
                        enum: [cache, upstream]
        '401':
          description: Missing or invalid credential
        '403':
          description: Company inactive or key revoked
        '429':
          description: Rate limit exceeded
        '503':
          description: Remote catalog unavailable and no valid cache exists
  /api/v1/my-inventory:
    get:
      summary: List the authenticated partner inventory
      tags:
        - Partner Inventory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Effective partner inventory
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartnerInventoryItem'
                  meta:
                    type: object
                    required:
                      - count
                      - companyId
                    properties:
                      count:
                        type: integer
                      companyId:
                        type: string
        '401':
          description: Missing or invalid credential
        '403':
          description: Company inactive or key revoked
        '429':
          description: Rate limit exceeded
  /api/v1/my-inventory/{productId}:
    patch:
      summary: Update the authenticated partner stock for one product
      tags:
        - Partner Inventory
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - custom_stock_quantity
              properties:
                custom_stock_quantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Partner stock updated
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/PartnerInventoryItem'
        '401':
          description: Missing or invalid credential
        '403':
          description: Company inactive or key revoked
        '404':
          description: Product not found in the authenticated partner inventory
  /api/internal/webhooks/supabase-sync:
    post:
      summary: Synchronize the local master catalog from Supabase
      tags:
        - Internal Webhooks
      security:
        - webhookSecret: []
      responses:
        '200':
          description: Master catalog synchronized
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - syncedCount
                      - updatedAt
                    properties:
                      syncedCount:
                        type: integer
                      updatedAt:
                        type: string
                        format: date-time
        '401':
          description: Invalid webhook secret
  /api/internal/admin/companies:
    get:
      summary: List companies for the Super Admin dashboard
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '200':
          description: Company list with API key counters
    post:
      summary: Create a company from the Super Admin dashboard
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '201':
          description: Company created
  /api/internal/admin/cost-settings:
    get:
      summary: Read the persisted cost settings used in product pricing
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '200':
          description: Current cost settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/CostSettings'
  /api/internal/admin/cost-settings/history:
    get:
      summary: List the full history of persisted cost setting changes
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '200':
          description: Cost setting history
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CostSettingsHistoryEntry'
    patch:
      summary: Update the persisted cost settings used by the API cost calculation
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '200':
          description: Updated cost settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/CostSettings'
  /api/internal/admin/companies/{companyId}:
    patch:
      summary: Update the selected company settings
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company updated
        '404':
          description: Company not found
  /api/internal/admin/companies/{companyId}/api-keys:
    get:
      summary: List API keys for a specific company
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key list
  /api/internal/admin/companies/{companyId}/status:
    patch:
      summary: Activate or deactivate a company from the internal admin panel
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company status updated
  /api/internal/admin/companies/{companyId}/inventory:
    get:
      summary: List the master catalog with the custom stock for one company
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company inventory for the Super Admin
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartnerInventoryItem'
                  meta:
                    type: object
                    required:
                      - count
                      - companyId
                      - companyName
                    properties:
                      count:
                        type: integer
                      companyId:
                        type: string
                      companyName:
                        type: string
  /api/internal/admin/companies/{companyId}/inventory/{productId}:
    put:
      summary: Update the custom stock for a company and product
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customStockQuantity
              properties:
                customStockQuantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Company inventory updated
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/PartnerInventoryItem'
        '404':
          description: Company or product not found
  /api/internal/admin/api-keys:
    post:
      summary: Issue a new API key from the internal admin panel
      tags:
        - Internal Admin
      security:
        - adminToken: []
      responses:
        '201':
          description: API key issued
  /api/internal/admin/api-keys/{apiKeyId}/revoke:
    patch:
      summary: Revoke an API key from the internal admin panel
      tags:
        - Internal Admin
      security:
        - adminToken: []
      parameters:
        - in: path
          name: apiKeyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key revoked
  /admin/companies:
    post:
      summary: Create a company
      tags:
        - Admin
      responses:
        '201':
          description: Company created
    get:
      summary: List companies
      tags:
        - Admin
      responses:
        '200':
          description: Company list
  /admin/companies/{companyId}/status:
    patch:
      summary: Activate or deactivate a company
      tags:
        - Admin
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company status updated
  /admin/api-keys:
    post:
      summary: Issue a new API key
      tags:
        - Admin
      responses:
        '201':
          description: API key issued
  /admin/api-keys/{apiKeyId}/revoke:
    patch:
      summary: Revoke an API key
      tags:
        - Admin
      parameters:
        - in: path
          name: apiKeyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key revoked
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
  schemas:
    Product:
      type: object
      required:
        - id
        - sku
        - name
        - serialNumber
        - availableQuantity
      properties:
        id:
          type: string
        sku:
          type: string
        numero_serie:
          type: string
        name:
          type: string
        nome:
          type: string
        serialNumber:
          type: string
        description:
          type: string
          nullable: true
        descricao:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        categoria:
          type: string
          nullable: true
        subcategory:
          type: string
          nullable: true
        subcategoria:
          type: string
          nullable: true
        baseMaterial:
          type: string
          nullable: true
        material_base:
          type: string
          nullable: true
        purity:
          type: string
          nullable: true
        pureza:
          type: string
          nullable: true
        weightGrams:
          oneOf:
            - type: string
            - type: number
          nullable: true
        peso_gramas:
          oneOf:
            - type: string
            - type: number
          nullable: true
        bathType:
          type: string
          nullable: true
        tipo_banho:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        bronzeImageKey:
          type: string
          nullable: true
        s3_key_bronze:
          type: string
          nullable: true
        silverImageKey:
          type: string
          nullable: true
        s3_key_silver:
          type: string
          nullable: true
        supplierCode:
          type: string
          nullable: true
        supplier_code:
          type: string
          nullable: true
        fiscalCode:
          type: string
          nullable: true
        fiscal_code:
          type: string
          nullable: true
        categoryId:
          type: string
          nullable: true
        category_id:
          type: string
          nullable: true
        productType:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
        typeId:
          type: string
          nullable: true
        type_id:
          type: string
          nullable: true
        subcategoryId:
          type: string
          nullable: true
        subcategory_id:
          type: string
          nullable: true
        blingProductId:
          type: string
          nullable: true
        bling_product_id:
          type: string
          nullable: true
        blingLastSyncAt:
          type: string
          format: date-time
          nullable: true
        bling_last_sync_at:
          type: string
          format: date-time
          nullable: true
        laborRateId:
          type: string
          nullable: true
        labor_rate_id:
          type: string
          nullable: true
        laborRateLabel:
          type: string
          nullable: true
        labor_rate_label:
          type: string
          nullable: true
        laborCost:
          oneOf:
            - type: string
            - type: number
          nullable: true
        labor_cost:
          oneOf:
            - type: string
            - type: number
          nullable: true
        sizeOptionId:
          type: string
          nullable: true
        size_option_id:
          type: string
          nullable: true
        sizeLabel:
          type: string
          nullable: true
        size_label:
          type: string
          nullable: true
        colorOptionId:
          type: string
          nullable: true
        color_option_id:
          type: string
          nullable: true
        colorLabel:
          type: string
          nullable: true
        color_label:
          type: string
          nullable: true
        availableQuantity:
          type: number
        available_quantity:
          type: number
          nullable: true
        stock_quantity:
          type: number
          nullable: true
        ncm:
          type: string
          nullable: true
        laborRateTableId:
          type: string
          nullable: true
        labor_rate_table_id:
          type: string
          nullable: true
        laborRateTableName:
          type: string
          nullable: true
        labor_rate_table_name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        price:
          type: number
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        costFinal:
          type: number
        costBreakdown:
          $ref: '#/components/schemas/ProductCostBreakdown'
    ProductCostBreakdown:
      type: object
      required:
        - laborCostUsd
        - laborCostBrl
        - silverCost
        - r1
        - r2
        - r3
        - finalCost
      properties:
        laborCostUsd:
          type: number
        laborCostBrl:
          type: number
        silverCost:
          type: number
        r1:
          type: number
        r2:
          type: number
        r3:
          type: number
        finalCost:
          type: number
    CostSettings:
      type: object
      required:
        - silverPricePerGram
        - zonaFrancaRatePercent
        - transportFee
        - dollarRate
        - updatedAt
      properties:
        silverPricePerGram:
          type: number
        zonaFrancaRatePercent:
          type: number
        transportFee:
          type: number
        dollarRate:
          type: number
        updatedAt:
          type: string
          format: date-time
    CostSettingsHistoryEntry:
      type: object
      required:
        - id
        - changedFields
        - previous
        - next
        - createdAt
      properties:
        id:
          type: string
        changedFields:
          type: array
          items:
            type: string
        previous:
          $ref: '#/components/schemas/CostSettingsHistorySnapshot'
        next:
          $ref: '#/components/schemas/CostSettingsHistorySnapshot'
        createdAt:
          type: string
          format: date-time
    CostSettingsHistorySnapshot:
      type: object
      required:
        - silverPricePerGram
        - zonaFrancaRatePercent
        - transportFee
        - dollarRate
      properties:
        silverPricePerGram:
          type: number
        zonaFrancaRatePercent:
          type: number
        transportFee:
          type: number
        dollarRate:
          type: number
    PartnerInventoryItem:
      type: object
      required:
        - productId
        - sku
        - name
        - masterStock
        - effectiveStockQuantity
        - updatedAt
      properties:
        productId:
          type: string
        sku:
          type: string
        name:
          type: string
        masterStock:
          type: integer
        customStockQuantity:
          type: integer
          nullable: true
        effectiveStockQuantity:
          type: integer
        updatedAt:
          type: string
          format: date-time